langtrace/requirements.txt
python-dotenv
opentelemetry-exporter-otlp-proto-http
langtrace_python_sdk
langchain
langchain_openai
langchain_community